refactor(web): extract tmux session lifecycle into agent-process
Move agent process management (start/stop/status/session detection) and
tmux send-keys/capture helpers (sendPromptToSession, capturePane,
isSessionReadyForPrompt) into src/web/agent-process.ts. web.ts shrinks
from 4371 to 4195 lines.
